Change the way the intrinsics size checking is done + mutable accessor for optimizers

+++ aslam_cv/include/aslam/cameras/distortion-radtan.h
@@ -113,7 +113,7 @@ class RadTanDistortion : public aslam::Distortion {
 
   /// \brief Returns the number of parameters used in the distortion model.
   ///        NOTE: Use the constexpr function parameterCount if you know the exact distortion type.
-  virtual size_t getParameterSize() const {
+  virtual int getParameterSize() const {
     return kNumOfParams;
   }
